+++ hw12_13_14_15_calendar/internal/logger/logger.go
@@ -2,12 +2,11 @@ package logger
 
 import (
 	"fmt"
+	"github.com/novikov-ai/practice-misc/hw12_13_14_15_calendar/configs"
 	"io"
 	"log"
 	"os"
 	"strings"
-
-	"github.com/novikov-ai/practice-misc/hw12_13_14_15_calendar/internal/configs"
 )
 
 type Logger struct {
